new hometask (decomposition)
