+++ 3-cli/storage/storage.go
@@ -1,0 +1,1 @@
+package storage
